Module icu_normalizer::properties

Access to the Unicode properties or property-based operations that are required for NFC and NFD.

Applications should generally use the full normalizers that are provided at the top level of this crate. However, the APIs in this module are provided for callers such as HarfBuzz that specifically want access to the raw canonical composition operation e.g. for use in a glyph-availability-guided custom normalizer.

Structs

CanonicalCombiningClassMap

Lookup of the Canonical_Combining_Class Unicode property.

CanonicalComposition

The raw canonical composition operation.

CanonicalDecomposition

The raw (non-recursive) canonical decomposition operation.

Enums

Decomposed

The outcome of non-recursive canonical decomposition of a character.