icu_normalizer::properties

Struct CanonicalComposition

pub struct CanonicalComposition { /* private fields */ }

The raw canonical composition operation.

Callers should generally use ComposingNormalizer instead of this API. However, this API is provided for callers such as HarfBuzz that specifically want access to the raw canonical composition operation e.g. for use in a glyph-availability-guided custom normalizer.

Implementations

impl CanonicalComposition

pub fn compose(&self, starter: char, second: char) -> Option<char>

Performs canonical composition (including Hangul) on a pair of characters or returns None if these characters donโt compose. Composition exclusions are taken into account.

Examples

let comp = icu::normalizer::properties::CanonicalComposition::new();

assert_eq!(comp.compose('a', 'b'), None); // Just two non-composing starters
assert_eq!(comp.compose('a', '\u{0308}'), Some('รค'));
assert_eq!(comp.compose('แบน', '\u{0302}'), Some('แป'));
assert_eq!(comp.compose('๐', '๐ฅ'), None); // Composition exclusion
assert_eq!(comp.compose('เง', 'เฆพ'), Some('เง')); // Second is starter
assert_eq!(comp.compose('แ', 'แก'), Some('๊ฐ')); // Hangul LV
assert_eq!(comp.compose('๊ฐ', 'แจ'), Some('๊ฐ')); // Hangul LVT

pub const fn new() -> Self

Constructs a new CanonicalComposition using compiled data.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

pub fn try_new_with_any_provider( provider: &(impl AnyProvider + ?Sized), ) -> Result<Self, NormalizerError>

A version of [Self :: new] that uses custom data provided by an AnyProvider.

๐ Help choosing a constructor

pub fn try_new_unstable<D>(provider: &D) -> Result<Self, NormalizerError>
where D: DataProvider<CanonicalCompositionsV1Marker> + ?Sized,

A version of Self::new that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

โ๏ธ The bounds on provider may change over time, including in SemVer minor releases.