icu_normalizer

Module provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Structs

Baked

Baked data

CanonicalCompositions

Non-Hangul canonical compositions

DecompositionData

Decomposition data

DecompositionTables

The expansion tables for cases where the decomposition isn’t contained in the trie value

NonRecursiveDecompositionSupplement

Non-recursive canonical decompositions that differ from DecompositionData.

NormalizerNfcV1

Marker for data for composition.

NormalizerNfdDataV1

Marker for data for canonical decomposition.

NormalizerNfdSupplementV1

Marker for additional data for non-recusrsive composition.

NormalizerNfdTablesV1

Marker for additional data for canonical decomposition.

NormalizerNfkdDataV1

Marker for data for compatibility decomposition.

NormalizerNfkdTablesV1

Marker for additional data for compatibility decomposition.

NormalizerUts46DataV1

Marker for data for UTS-46 decomposition.