icu_normalizer

Module provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Structs

Baked

Baked data

CanonicalCompositionsV1

Non-Hangul canonical compositions

CanonicalCompositionsV1Marker

Marker type for CanonicalCompositionsV1: “normalizer/comp@1”

CanonicalDecompositionDataV1Marker

Marker type for DecompositionDataV1: “normalizer/nfd@1”

CanonicalDecompositionTablesV1Marker

Marker type for DecompositionTablesV1: “normalizer/nfdex@1”

CompatibilityDecompositionSupplementV1Marker

Marker type for DecompositionSupplementV1: “normalizer/nfkd@1”

CompatibilityDecompositionTablesV1Marker

Marker type for DecompositionTablesV1: “normalizer/nfkdex@1”

DecompositionDataV1

Main data for NFD

DecompositionSupplementV1

Data that either NFKD or the decomposed form of UTS 46 needs in addition to the NFD data.

DecompositionTablesV1

The expansion tables for cases where the decomposition isn’t contained in the trie value

NonRecursiveDecompositionSupplementV1

Non-recursive canonical decompositions that differ from DecompositionDataV1.

NonRecursiveDecompositionSupplementV1Marker

Marker type for NonRecursiveDecompositionSupplementV1: “normalizer/decomp@1”

Uts46DecompositionSupplementV1Marker

Marker type for DecompositionSupplementV1: “normalizer/uts46d@1”